backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
pillow==10.1.0
python-dotenv==1.0.0